adnuts example for admb
"adnuts_eample_labmeeting.R" contains the code to run. mcpars.R, kobe.R, and read.admb.R are needed to create some of the plots, and the admb model files are pm.tpl, hakedata.dat, pm.ctl, and pm.dat. Everything should be in the same folder and the path to that folder needs to be set in "adnuts_eample_labmeeting.R"